# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/admin/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get tasks list",
                "description": "Get tasks list",
                "operationId": "97234e25b4c5f4fe69bb3e0b89793e73",
                "parameters": [
                    {
                        "name": "searchKey",
                        "in": "query",
                        "description": "parameter for search in title and explain"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 64,
                                                        "nullable": false
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1,
                                                        "nullable": false
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "تست جدید",
                                                        "nullable": false
                                                    },
                                                    "explain": {
                                                        "type": "string",
                                                        "example": "جالبه",
                                                        "nullable": true
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "running",
                                                            "done"
                                                        ],
                                                        "example": "running",
                                                        "nullable": false
                                                    },
                                                    "priority": {
                                                        "type": "string",
                                                        "enum": [
                                                            "low",
                                                            "medium",
                                                            "high"
                                                        ],
                                                        "example": "medium",
                                                        "nullable": false
                                                    },
                                                    "end_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "1403/06/01",
                                                        "nullable": false
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-22T15:03:17.000000Z",
                                                        "nullable": false
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-22T15:03:17.000000Z",
                                                        "nullable": false
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/v1/admin/tasks?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/v1/admin/tasks?page=1"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "example": null,
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "1"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tasks"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}